pyproject.toml
[tool.poetry]
name = "django-challenge-sc"
version = "0.2.0"
description = ""
authors = ["Malazay SouthernCode <alexisgiovoglanian@southerncode.us>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.1"
django-rest-framework = "^0.1.0"
pytest = "^7.2.0"
django-filter = "^22.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
